config with connector
