+++ src/main/java/fr/edf/jenkins/plugins/windows/WindowsHost.groovy
@@ -1,36 +1,16 @@
 package fr.edf.jenkins.plugins.windows
 
 
-import static com.cloudbees.plugins.credentials.CredentialsMatchers.anyOf
-import static com.cloudbees.plugins.credentials.CredentialsMatchers.instanceOf
-import static com.cloudbees.plugins.credentials.domains.URIRequirementBuilder.fromUri
-
 import org.apache.commons.lang.StringUtils
-import org.apache.http.client.config.AuthSchemes
-import org.kohsuke.stapler.AncestorInPath
 import org.kohsuke.stapler.DataBoundConstructor
 import org.kohsuke.stapler.DataBoundSetter
-import org.kohsuke.stapler.QueryParameter
-import org.kohsuke.stapler.verb.POST
-
-import com.cloudbees.plugins.credentials.CredentialsProvider
-import com.cloudbees.plugins.credentials.common.StandardCredentials
-import com.cloudbees.plugins.credentials.common.StandardListBoxModel
-import com.cloudbees.plugins.credentials.common.StandardUsernamePasswordCredentials
 
 import fr.edf.jenkins.plugins.windows.connector.WindowsComputerConnector
-import fr.edf.jenkins.plugins.windows.util.FormUtils
-import fr.edf.jenkins.plugins.windows.winrm.WinRMCommand
-import fr.edf.jenkins.plugins.windows.winrm.connection.WinRMGlobalConnectionConfiguration
 import hudson.Extension
 import hudson.model.Describable
 import hudson.model.Descriptor
-import hudson.model.Item
 import hudson.model.Label
 import hudson.model.labels.LabelAtom
-import hudson.security.ACL
-import hudson.util.FormValidation
-import hudson.util.ListBoxModel
 import jenkins.model.Jenkins
 /**
  * WinRm conncetion configuration for a Windows host
@@ -247,78 +227,6 @@ class WindowsHost implements Describable<WindowsHost> {
         @Override
         String getDisplayName() {
             return Messages.Host_DefaultName()
-        }
-
-        /**
-         * List the available authentication schemes for WinRm. However NTLM is the default one
-         * @param authenticationScheme
-         * @return authentication scheme
-         */
-        @POST
-        ListBoxModel doFillAuthenticationSchemeItems() {
-            ListBoxModel result = new ListBoxModel()
-            [AuthSchemes.NTLM, AuthSchemes.BASIC].each {
-                result.add(it,it)
-            }
-            return result
-        }
-
-        /**
-         * List the available credentials
-         * @param host
-         * @param credentialsId
-         * @param item
-         * @return CredentialsId
-         */
-        @POST
-        ListBoxModel doFillCredentialsIdItems(@QueryParameter String host, @QueryParameter String credentialsId,
-                @AncestorInPath Item item) {
-            StandardListBoxModel result = new StandardListBoxModel()
-            boolean notAdmin = item == null && !Jenkins.get().hasPermission(Jenkins.ADMINISTER)
-            boolean noCredentials = item != null && !item.hasPermission(Item.EXTENDED_READ) &&
-                    !item.hasPermission(CredentialsProvider.USE_ITEM)
-
-            if(notAdmin || noCredentials) {
-                return result.includeCurrentValue(credentialsId)
-            }
-            return result
-                    .includeEmptyValue()
-                    .includeMatchingAs(ACL.SYSTEM,
-                    item ?: Jenkins.get(),
-                    StandardCredentials.class,
-                    fromUri(FormUtils.getUri(host).toString()).build(),
-                    anyOf(instanceOf(StandardUsernamePasswordCredentials)))
-                    .includeCurrentValue(credentialsId)
-        }
-
-        /**
-         * Checks connection on Windows machine
-         * @param host
-         * @param port
-         * @param credentialsId
-         * @param authenticationScheme
-         * @param useHttps
-         * @param item
-         * @return connection success otherwise connection failed
-         */
-        @POST
-        FormValidation doVerifyConnection(@QueryParameter String host, @QueryParameter Integer port,
-                @QueryParameter String credentialsId, @QueryParameter String authenticationScheme,
-                @QueryParameter Boolean useHttps, @QueryParameter Boolean disableCertificateCheck,
-                @QueryParameter Integer connectionTimeout, @QueryParameter Integer readTimeout,
-                @AncestorInPath Item item) {
-
-            try {
-
-                Jenkins.get().checkPermission(Jenkins.ADMINISTER)
-                String result = WinRMCommand.checkConnection(new WinRMGlobalConnectionConfiguration(
-                        credentialsId: credentialsId, context: item, host: host, port: port, authenticationScheme: authenticationScheme,
-                        useHttps: useHttps, disableCertificateCheck: disableCertificateCheck,
-                        connectionTimeout: connectionTimeout, readTimeout: readTimeout))
-                return FormValidation.ok("Connection success : " + (result).toString())
-            } catch(Exception e) {
-                return FormValidation.error("Connection failed : " + (e.getMessage()).toString())
-            }
         }
     }
 }
