+++ src/main/java/fr/edf/jenkins/plugins/windows/connector/MicroServiceJNLPConnector.groovy
@@ -3,29 +3,60 @@ package fr.edf.jenkins.plugins.windows.connector
 import java.time.Instant
 
 import org.apache.commons.lang.exception.ExceptionUtils
+import org.jenkinsci.Symbol
 
 import fr.edf.jenkins.plugins.windows.WindowsHost
 import fr.edf.jenkins.plugins.windows.WindowsUser
 import fr.edf.jenkins.plugins.windows.agent.WindowsComputer
+import fr.edf.jenkins.plugins.windows.winrm.WinRMCommandException
+import hudson.Extension
+import hudson.model.Descriptor
 import hudson.model.TaskListener
 import hudson.slaves.ComputerLauncher
 import hudson.slaves.JNLPLauncher
 import hudson.slaves.SlaveComputer
 
 class MicroServiceJNLPConnector extends WindowsComputerConnector {
-    
+
     @Override
     protected ComputerLauncher createLauncher(WindowsHost host, WindowsUser user) {
-        return new WindowsJNLPLauncher(host, user, jenkinsUrl)
+        return new MicroServiceJNLPLauncher(host, user, jenkinsUrl)
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    protected List<String> listUsers(WindowsHost host) throws WinRMCommandException {
+        return null;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    protected void deleteUser(WindowsHost host, String username) throws WinRMCommandException, Exception {
     }
     
-    private static class WinRmJNLPLauncher extends JNLPLauncher {
+    @Extension @Symbol("microservice")
+    static final class DescriptorImpl extends Descriptor<WindowsComputerConnector>{
+
+        /**
+         * {@inheritDoc}
+         */
+        @Override
+        String getDisplayName() {
+            return "Connect with Microservice and JNLP"
+        }
+    }
+
+    private static class MicroServiceJNLPLauncher extends JNLPLauncher {
         WindowsHost host
         WindowsUser user
         String jenkinsUrl
         boolean launched
 
-        WinRmJNLPLauncher(WindowsHost host, WindowsUser user, String jenkinsUrl) {
+        MicroServiceJNLPLauncher(WindowsHost host, WindowsUser user, String jenkinsUrl) {
             super(true)
             this.host = host
             this.user = user
@@ -48,7 +79,6 @@ class MicroServiceJNLPConnector extends WindowsComputerConnector {
             launched = true
             WindowsComputer windowsComputer = (WindowsComputer) computer
             try {
-                
             }catch(Exception e) {
                 launched = false
                 String message = String.format("Error while connecting computer %s due to %s ",
