+++ src/main/resources/fr/edf/jenkins/plugins/windows/WindowsCloud/config.groovy
@@ -22,6 +22,5 @@ f.advanced(title: Messages.Cloud_Details()) {
         f.entry(title:Messages.Cloud_IdleMinutes(), field:'idleMinutes') {
             f.number(clazz: 'required', min: 1, default: 1)
         }
-        f.dropdownDescriptorSelector(title:'Connect method', field:'connector')
     }
 }
