+++ src/main/resources/fr/edf/jenkins/plugins/windows/connector/WinRmJNLPConnector/help-maxTries.html
