--- src/main/resources/fr/edf/jenkins/plugins/windows/connector/WindowsComputerJNLPConnector/config.groovy
@@ -1,7 +1,0 @@
-import fr.edf.jenkins.plugins.windows.Messages
-
-def f = namespace(lib.FormTagLib)
-
-f.entry(title: Messages._Cloud_JenkinsUrl(), field: 'jenkinsUrl'){
-    f.textbox()
-}
